Disable qtmultimedia tests requiring Windows Media Foundation
The WMF was introduced in Windows Vista and is not supported on Windows XP in
the scope that qtmultimedia requires. However, because we're effectively using
the Windows 7 SDK even when building for Windows XP, these configuration tests
pass and produce libraries with dependencies to MF.DLL which is not present on
Windows XP. The best way out is thus to simply make them error out
unconditionally, which causes the configure script to skip these submodules.